+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Run the google covid pipeline daily 
 - Allow null columns on the derived export wins pipeline
 - Use the high memory queue for the google covid dataset fetch operation
+- Add region name and region code fields to oxford covid dataset
 
 ## 2020-09-18